Move create_session_cookie to utils.py
It isn't used by anything in auth, only in tests.
